# src/Gatherstead.Api/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Gatherstead.Api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/tenants/{tenantId}/accommodation-intents": {
      "get": {
        "tags": [
          "AccommodationIntentsQuery"
        ],
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "memberIds",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fromNight",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MyStayDtoIReadOnlyCollectionBaseEntityResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyStayDtoIReadOnlyCollectionBaseEntityResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyStayDtoIReadOnlyCollectionBaseEntityResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tenants/{tenantId}/task-intents": {
      "get": {
        "tags": [
          "TaskIntentsQuery"
        ],
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "memberIds",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fromDay",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MyTaskDtoIReadOnlyCollectionBaseEntityResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyTaskDtoIReadOnlyCollectionBaseEntityResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyTaskDtoIReadOnlyCollectionBaseEntityResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MyStayDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "accommodationId": {
            "type": "string",
            "format": "uuid"
          },
          "accommodationName": {
            "type": "string"
          },
          "propertyId": {
            "type": "string",
            "format": "uuid"
          },
          "propertyName": {
            "type": "string"
          },
          "householdMemberId": {
            "type": "string",
            "format": "uuid"
          },
          "startNight": {
            "type": "string",
            "format": "date"
          },
          "endNight": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "enum": [
              "Intent",
              "Hold",
              "Confirmed"
            ],
            "type": "string"
          },
          "decision": {
            "enum": [
              "Pending",
              "Approved",
              "Declined"
            ],
            "type": "string"
          },
          "partyAdults": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "partyChildren": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MyStayDtoIReadOnlyCollectionBaseEntityResponse": {
        "type": "object",
        "properties": {
          "entity": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MyStayDto"
            },
            "nullable": true
          },
          "successful": {
            "type": "boolean"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResponseMessage"
            },
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "MyTaskDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "taskPlanId": {
            "type": "string",
            "format": "uuid"
          },
          "householdMemberId": {
            "type": "string",
            "format": "uuid"
          },
          "taskName": {
            "type": "string"
          },
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "eventName": {
            "type": "string"
          },
          "day": {
            "type": "string",
            "format": "date"
          },
          "timeSlot": {
            "enum": [
              "Morning",
              "Midday",
              "Evening",
              "Anytime"
            ],
            "type": "string",
            "nullable": true
          },
          "completed": {
            "type": "boolean"
          },
          "volunteered": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "MyTaskDtoIReadOnlyCollectionBaseEntityResponse": {
        "type": "object",
        "properties": {
          "entity": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MyTaskDto"
            },
            "nullable": true
          },
          "successful": {
            "type": "boolean"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResponseMessage"
            },
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "ResponseMessage": {
        "type": "object",
        "properties": {
          "type": {
            "enum": [
              "ERROR",
              "WARNING",
              "INFORMATION"
            ],
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}